Fixed API test for downloading binary data.
Also improved the Travis CI container
framework so that we can skip tests or
provide filters and verbose levels for
the API tests. All this is done with
environment variables.

+++ Travis/cb_run_tests.sh
@@ -152,6 +152,24 @@ fail_bourreau=""
 fail_api_curl=""
 fail_api_ruby=""
 
+# -------------------------------
+# Show TEST environment variables
+# -------------------------------
+# These environment variable allow the
+# user to skip over some test stages,
+# or make the API tests more verbose.
+printf "${BLUE}Environment variables for this test session:${NC}\n"
+echo   ""
+echo   "General control:"
+printf "${YELLOW}CBRAIN_SKIP_TEST_STAGES${NC}        = '${CBRAIN_SKIP_TEST_STAGES:=unset}'\n"
+echo   ""
+echo   "API test control:"
+printf "${YELLOW}CBRAIN_CURL_TEST_VERBOSE_LEVEL${NC} = '${CBRAIN_CURL_TEST_VERBOSE_LEVEL:=1}'\n"
+printf "${YELLOW}CBRAIN_CURL_TEST_FILTER${NC}        = '${CBRAIN_CURL_TEST_FILTER}'\n"
+printf "${YELLOW}CBRAIN_GEM_TEST_VERBOSE_LEVEL${NC}  = '${CBRAIN_GEM_TEST_VERBOSE_LEVEL:=1}'\n"
+printf "${YELLOW}CBRAIN_GEM_TEST_FILTER${NC}         = '${CBRAIN_GEM_TEST_FILTER}'\n"
+echo   ""
+
 # ------------------------------
 # Portal-Side Testing
 # ------------------------------
@@ -160,7 +178,11 @@ cd $cb_test/BrainPortal || die "Cannot cd to BrainPortal directory"
 # Eventually, it would be nice if from a ENV variable set in Travis,
 # we could run only a subset of the tests.
 printf "${BLUE}Running rspec on BrainPortal side.${NC}\n"
-rspec spec || fail_portal="rspec on BrainPortal failed with return code $?"
+if echo "X$CBRAIN_SKIP_TEST_STAGES" | grep -q 'RspecPortal' >/dev/null ; then
+  printf "${YELLOW} -> Skipped by request from env CBRAIN_SKIP_TEST_STAGES${NC}\n"
+else
+  rspec spec || fail_portal="rspec on BrainPortal failed with return code $?"
+fi
 #CBRAIN_FAILTEST=1 rspec spec/modules/travis_ci_spec.rb || fail_portal="rspec on BrainPortal failed with return code $?"
 
 
@@ -175,7 +197,11 @@ cd $cb_test/Bourreau || die "Cannot cd to Bourreau directory"
 # -> NOTE FIXME TODO : hardcoded 'spec/boutiques' for <-
 # -> the moment because no other test files work on Bourreau. <-
 printf "${BLUE}Running rspec on Bourreau side.${NC}\n"
-rspec spec/boutiques || fail_bourreau="rspec on Bourreau failed with return code $?"
+if echo "X$CBRAIN_SKIP_TEST_STAGES" | grep -q 'RspecBourreau' >/dev/null ; then
+  printf "${YELLOW} -> Skipped by request from env CBRAIN_SKIP_TEST_STAGES${NC}\n"
+else
+  rspec spec/boutiques || fail_bourreau="rspec on Bourreau failed with return code $?"
+fi
 
 
 
@@ -183,27 +209,45 @@ rspec spec/boutiques || fail_bourreau="rspec on Bourreau failed with return code
 # Testing of API (curl)
 # ------------------------------
 printf "${BLUE}Running API tests with curl.${NC}\n"
-cd $cb_test/BrainPortal            || die "Cannot cd to BrainPortal directory"
-rake "db:seed:test:api" >/dev/null || die "Cannot re-seed the DB for API testing"
-rails server puma -p 3000 -d       || die "Cannot start local puma server?"
-cd test_api                        || die "Cannot cd to test_api directory?"
-sleep 5 # must wait a bit for puma to be ready
-perl curl_req_tester.pl -h localhost -p 3000 -s http -R || fail_api_curl="API testing with CURL failed"
-kill $(cat $cb_test/BrainPortal/tmp/pids/server.pid)
+if echo "X$CBRAIN_SKIP_TEST_STAGES" | grep -q 'CurlAPI' >/dev/null ; then
+  printf "${YELLOW} -> Skipped by request from env CBRAIN_SKIP_TEST_STAGES${NC}\n"
+else
+  cd $cb_test/BrainPortal            || die "Cannot cd to BrainPortal directory"
+  rake "db:seed:test:api" >/dev/null || die "Cannot re-seed the DB for API testing"
+  rails server puma -p 3000 -d       || die "Cannot start local puma server?"
+  cd test_api                        || die "Cannot cd to test_api directory?"
+  sleep 5 # must wait a bit for puma to be ready
+  perl curl_req_tester.pl                 \
+    -h localhost                          \
+    -p 3000                               \
+    -s http                               \
+    -v"${CBRAIN_CURL_TEST_VERBOSE_LEVEL}" \
+    -R                                    \
+    ${CBRAIN_CURL_TEST_FILTER}            \
+    || fail_api_curl="API testing with CURL failed"
+  kill $(cat $cb_test/BrainPortal/tmp/pids/server.pid)
+fi
 
 
 
 # ------------------------------
-# Testing of API (Ruby)
+# Testing of API (Ruby Gem)
 # ------------------------------
 printf "${BLUE}Running API tests with Ruby CbrainClient gem.${NC}\n"
-cd $cb_test/BrainPortal            || die "Cannot cd to BrainPortal directory"
-rake "db:seed:test:api" >/dev/null || die "Cannot re-seed the DB for API testing"
-rails server puma -p 3000 -d       || die "Cannot start local puma server?"
-cd test_api                        || die "Cannot cd to test_api directory?"
-sleep 5 # must wait a bit for puma to be ready
-rake "cbrain:test:api:client" || fail_api_ruby="API testing with Ruby CbrainClient failed"
-kill $(cat $cb_test/BrainPortal/tmp/pids/server.pid)
+if echo "X$CBRAIN_SKIP_TEST_STAGES" | grep -q 'GemAPI' >/dev/null ; then
+  printf "${YELLOW} -> Skipped by request from env CBRAIN_SKIP_TEST_STAGES${NC}\n"
+else
+  cd $cb_test/BrainPortal            || die "Cannot cd to BrainPortal directory"
+  rake "db:seed:test:api" >/dev/null || die "Cannot re-seed the DB for API testing"
+  rails server puma -p 3000 -d       || die "Cannot start local puma server?"
+  cd test_api                        || die "Cannot cd to test_api directory?"
+  sleep 5 # must wait a bit for puma to be ready
+  rake "cbrain:test:api:client"           \
+    -v "${CBRAIN_GEM_TEST_VERBOSE_LEVEL}" \
+    ${CBRAIN_GEM_TEST_FILTER}             \
+    || fail_api_ruby="API testing with Ruby CbrainClient failed"
+  kill $(cat $cb_test/BrainPortal/tmp/pids/server.pid)
+fi
